Made it so that transfers are grouped and an be viewed in groups
The group is defined using a key defined inside the  watchdirs.yaml.
That means the database of progress for tasks is now  dictionary with a groupName key containing a list of transferData objects.
Accessing a cancellation token is done using the exact same indexing: groupName, index.

on the client side rending is done exactly as before except that there is now another lay above the previous renderer that handles providing lists of transferData objects to the previous rendering component


Former-commit-id: 4248039d5d2342dc70efdcb976c48b7bc9f8d7db