Moved and siginificantly refactored TransferProgress callbacks.
The Mover no longer has an eventhandler passed into it and that has been completely removed from all calling code.
It can be added in later and is simply not necessary for now plus it complicates the process of writing to the database.... though if i wanted to add it in i could have some converter that converts each progress type into one universal one that would allow much easier handling.
Most of the code for the two callbacks was able to be merged and is now in a seperate function
the Mover now gets a discriminated union containing the appropriate callback and then runs the specified move function for that callback


Former-commit-id: e9c236a509950730422078850463ca52df1b0c50