fixed a few silly bugs and made streams and FTPclients dispose correctly
watcher:
ftpclient is now sometimes null instead of option. This means we can have "use" instead of a let and it will automatically dispose
scheduler:
The filepath now has information about the file if it is ftp(size ). This means the scheduler doesn't have to make a connection to the ftp server
mover:
stopped accidently using ust the diretory instead of the filepath when getting the source stream. Made sure all streams disposed or are defined with a "use"

Former-commit-id: 6f28bb79f6bcf6fccad619c558e32dfaf1522d8e