Overhauled syncing code to no rwutire locks.
Instead of writing to a shared peice of data and having to lock it the vrious syncing commands go into an event stream.
When an updated transferdata object is made the event is calle and given the new data.
At an interval the event stream has all its data read and then each chane is applied sequentially to build a new UIDATA object which is then sent off.
this means there is no mutability and everything is thread safe and orderly. very very nice


Former-commit-id: 25d78ae683bd8c760fb81101936afb08220854c3