Smaller config and ui chnages
Config:
MaxJobs-> GroupMaxJobs
Added TotalMaxJobs to represent the limit for the entire client. now we don't have to manually specify a "main" category on every client
Code to make the client limit group be added to all watchDirs
Made it so the ui does not display that first group.
Added "DisplayPriorty" option for client config.
This sets the order in which the clients will show in the navbar in the gui