Fixed a major bug in cancellation and failure
It caused cancellation and failure to lock the job queue  becuase the cancel function was going through the request handler and wouldn't return untill the cancellation was complete. The completion of the cancellation would require a request handler request to be run. Whcihc was busy, handling the request for cancellation.