Made many of the DBRequests Async
This has given us a large performance uplift and allso exposed a large bug in the Access Code
The MakeJobFInished function in the access code was actually being partially apllied. What this meant was that instead of being scheduled to run the function on the rquesthandler the request handler would schedule a function that retuned a partial function which would then be applied and the actual MakeJobFinished func would run outside of the request handler.
Allos made many small improovmen'ts to logging and made all functions that can be private, private